Moving service binding workflows from test fixtures to conditionally loaded core classes
CreateBindingWithServiceKeyComponentTest Test OK
