+++ spring-cloud-app-broker-core/src/main/java/org/springframework/cloud/appbroker/workflow/ServiceKeyCreateServiceBindingWorkflow.java
@@ -14,77 +14,81 @@
  * limitations under the License.
  */
 
-package org.springframework.cloud.appbroker.integration.fixtures;
+package org.springframework.cloud.appbroker.workflow;
 
 import java.util.List;
+import java.util.Map;
+
+import org.springframework.cloud.appbroker.deployer.BackingService;
+import org.springframework.cloud.appbroker.deployer.BackingServiceKey;
+import org.springframework.cloud.appbroker.deployer.BackingServiceKeys;
+import org.springframework.cloud.appbroker.deployer.BackingServicesProvisionService;
+import org.springframework.cloud.appbroker.deployer.BrokeredServices;
+import org.springframework.cloud.appbroker.extensions.parameters.BackingServicesParametersTransformationService;
+import org.springframework.cloud.appbroker.extensions.targets.TargetService;
+import org.springframework.cloud.appbroker.service.CreateServiceInstanceAppBindingWorkflow;
+import org.springframework.cloud.servicebroker.model.binding.CreateServiceInstanceAppBindingResponse;
+import org.springframework.cloud.servicebroker.model.binding.CreateServiceInstanceBindingRequest;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.util.Logger;
 import reactor.util.Loggers;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.test.context.TestComponent;
-import org.springframework.cloud.appbroker.deployer.BackingService;
-import org.springframework.cloud.appbroker.deployer.BackingServiceKey;
-import org.springframework.cloud.appbroker.deployer.BackingServiceKeys;
-import org.springframework.cloud.appbroker.deployer.BackingServicesProvisionService;
-import org.springframework.cloud.appbroker.deployer.BrokeredServices;
-import org.springframework.cloud.appbroker.extensions.targets.TargetService;
-import org.springframework.cloud.appbroker.service.DeleteServiceInstanceBindingWorkflow;
-import org.springframework.cloud.servicebroker.model.binding.DeleteServiceInstanceBindingRequest;
-import org.springframework.cloud.servicebroker.model.binding.DeleteServiceInstanceBindingResponse;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
-
 @SuppressWarnings("WeakerAccess")
-@TestComponent
-@Order(Ordered.HIGHEST_PRECEDENCE)
-@ConditionalOnProperty(name="service-bindings-as-service-keys")
-public class ServiceKeyDeleteServiceBindingWorkflow
+public class ServiceKeyCreateServiceBindingWorkflow
 	extends AbstractServiceInstanceWorkflow
-	implements DeleteServiceInstanceBindingWorkflow {
+	implements CreateServiceInstanceAppBindingWorkflow {
 
-	private final Logger log = Loggers.getLogger(ServiceKeyDeleteServiceBindingWorkflow.class);
+	private final Logger log = Loggers.getLogger(ServiceKeyCreateServiceBindingWorkflow.class);
 
 	private final BackingServicesProvisionService backingServicesProvisionService;
 
+	private final BackingServicesParametersTransformationService servicesParametersTransformationService;
+
 	private final TargetService targetService;
 
-	public ServiceKeyDeleteServiceBindingWorkflow(BrokeredServices brokeredServices,
+	public ServiceKeyCreateServiceBindingWorkflow(BrokeredServices brokeredServices,
 		BackingServicesProvisionService backingServicesProvisionService,
+		BackingServicesParametersTransformationService servicesParametersTransformationService,
 		TargetService targetService) {
 		super(brokeredServices);
 		this.backingServicesProvisionService = backingServicesProvisionService;
+		this.servicesParametersTransformationService = servicesParametersTransformationService;
 		this.targetService = targetService;
 	}
 
 	@Override
-	public Mono<Boolean> accept(DeleteServiceInstanceBindingRequest request) {
+	public Mono<Boolean> accept(CreateServiceInstanceBindingRequest request) {
 		//Only accept binding request matching one registered backing service
 		return accept(request.getServiceDefinition(), request.getPlan());
 	}
 
 	@Override
-	public Mono<DeleteServiceInstanceBindingResponse.DeleteServiceInstanceBindingResponseBuilder> buildResponse(DeleteServiceInstanceBindingRequest request, DeleteServiceInstanceBindingResponse.DeleteServiceInstanceBindingResponseBuilder responseBuilder) {
-		return deleteBackingServiceKey(request).
-			then(Mono.just(responseBuilder));
+	public Mono<CreateServiceInstanceAppBindingResponse.CreateServiceInstanceAppBindingResponseBuilder> buildResponse(CreateServiceInstanceBindingRequest request, CreateServiceInstanceAppBindingResponse.CreateServiceInstanceAppBindingResponseBuilder responseBuilder) {
+
+		return createBackingServiceKey(request).   //create service key, returning credentials
+			last(). //only keep last one (should never happen)
+			map(responseBuilder::credentials); //store it in builder
 	}
 
-	private Flux<String> deleteBackingServiceKey(DeleteServiceInstanceBindingRequest request) {
+	private Flux<Map<String, Object>> createBackingServiceKey(CreateServiceInstanceBindingRequest request) {
 		return getBackingServicesForService(request.getServiceDefinition(), request.getPlan())
 			.flatMap(backingServices ->
 				targetService.addToBackingServices(backingServices,
 					getTargetForService(request.getServiceDefinition(), request.getPlan()) ,
 					request.getServiceInstanceId()))
-			.flatMap(ServiceKeyDeleteServiceBindingWorkflow::getBackingServiceKeys)
+			.flatMap(backingServices ->
+				servicesParametersTransformationService.transformParameters(backingServices,
+					request.getParameters()))
+			.flatMap(ServiceKeyCreateServiceBindingWorkflow::getBackingServiceKeys)
 			.flatMap(backingServiceKeys -> setServiceKeyName(backingServiceKeys, request.getBindingId()))
-			.flatMapMany(backingServicesProvisionService::deleteServiceKeys)
-			.doOnRequest(l -> log.debug("Deleting backing service keys for {}/{}",
+			.flatMapMany(backingServicesProvisionService::createServiceKeys)
+			.doOnRequest(l -> log.debug("Creating backing service keys for {}/{}",
 				request.getServiceDefinition().getName(), request.getPlan().getName()))
-			.doOnComplete(() -> log.debug("Finished deleting backing service keys for {}/{}",
+			.doOnComplete(() -> log.debug("Finished creating backing service keys for {}/{}",
 				request.getServiceDefinition().getName(), request.getPlan().getName()))
-			.doOnError(exception -> log.error("Error deleting backing services keysfor {}/{} with error '{}'",
+			.doOnError(exception -> log.error("Error creating backing services keysfor {}/{} with error '{}'",
 				request.getServiceDefinition().getName(), request.getPlan().getName(), exceptionMessageOrToString(exception)));
 	}
