+++ spring-cloud-app-broker-integration-tests/src/test/java/org/springframework/cloud/appbroker/integration/WiremockComponentTest.java
@@ -27,8 +27,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.cloud.appbroker.integration.fixtures.CloudControllerStubFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.CredHubStubFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.OpenServiceBrokerApiFixture;
-import org.springframework.cloud.appbroker.integration.fixtures.ServiceKeyCreateServiceBindingWorkflow;
-import org.springframework.cloud.appbroker.integration.fixtures.ServiceKeyDeleteServiceBindingWorkflow;
+import org.springframework.cloud.appbroker.integration.fixtures.ServiceKeyWorkflowsConfig;
 import org.springframework.cloud.appbroker.integration.fixtures.TestBindingCredentialsProviderFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.UaaStubFixture;
 import org.springframework.cloud.appbroker.integration.fixtures.WiremockServerFixture;
@@ -44,10 +43,8 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 		CloudControllerStubFixture.class,
 		UaaStubFixture.class,
 		CredHubStubFixture.class,
-		TestBindingCredentialsProviderFixture.class
-		,
-		ServiceKeyCreateServiceBindingWorkflow.class,
-		ServiceKeyDeleteServiceBindingWorkflow.class
+		TestBindingCredentialsProviderFixture.class,
+		ServiceKeyWorkflowsConfig.class
 	},
 	properties = {
 		"spring.cloud.appbroker.deployer.cloudfoundry.api-host=localhost",
